[#Ente-124] Moved IJiraConfig, jiraEmail declared as EmailAddress
